tetsing

+++ include/teb_local_planner/timed_elastic_band.hpp
@@ -150,10 +150,10 @@ bool TimedElasticBand::initTEBtoGoal(BidirIter path_start, BidirIter path_end, F
       PoseSE2 goal(goal_position, goal_orient);
 
       // if number of samples is not larger than min_samples, insert manually
-      if ( sizePoses() < min_samples-1 )
+      if ( (int)sizePoses() < min_samples-1 )
       {
         ROS_DEBUG("initTEBtoGoal(): number of generated samples is less than specified by min_samples. Forcing the insertion of more samples...");
-        while (sizePoses() < min_samples-1) // subtract goal point that will be added later
+        while ((int)sizePoses() < min_samples-1) // subtract goal point that will be added later
         {
           // simple strategy: interpolate between the current pose and the goal
           addPoseAndTimeDiff( PoseSE2::average(BackPose(), goal), timestep ); // let the optimier correct the timestep (TODO: better initialization
